Rename Block component to Blocks and deprecate an old
